move context decl to proper subpackage
